CW Issue #839: Move registry widgets to a separate composite
